Create lib.rs
Donation Smart Contract : 
Deposit: The deposit function allows the donor to send tokens to the smart contract. This is achieved through a transfer of tokens from the donor’s account to the donation account.

Withdraw: The withdraw function allows the donor to withdraw their tokens from the smart contract. The donor specifies the amount to withdraw.

BalanceOf: The balance_of function returns the balance of a particular user’s token account.

TotalBalance: The total_balance function returns the total balance of the donation pool by querying the contract's token account.